258.1. Displaying the User Header(Section 19: Redux Store Design)
Now that we've got our fette user action creator put together we're going to start to create our component

that is going to display just the author's name.

And again we're going to called out the user Hetter.

So back inside of my components directory I will make a new file called user header.

Yes.

And then inside of here we're going to put together some boilerplate for a class based component.

So I'm going to import re-act from react.

I'll create class User Hetter that extends re-act component and then inside of here I will define my

render method inside the render method.

Right now we'll just return a simple div that has a text like user Hetter And then as usual at the bottom

we will place our export default user header.

All right.

So we're going to want to make sure that this thing eventually is going to call an action creator to

fetch the user that it needs to show on the screen.

We'll get around to that in just a second.

But for now let's make sure that the user Hetter gets displayed inside of our posts list at the appropriate

location.

So going to flip back over to my posts list component at the top I'm going to import user Hetter rom

user Hetter and then down inside of my render list method we're going to create a new instance of this

component right after the closing div on the H-2 and that paragraph tax.

So right you're right underneath one div I'm going to put user Hetter.

Now when we create this component we need to tell it exactly what user id it needs to show.

Or excuse me the idea of the user that it needs to show.

So we need to pass that down as a prop.

Now we could certainly pass down a prop called ID.

But just to make the terminology inside of you're a little bit easier to follow I'm going to call this

Propp user id just make sure it's really clear.

We are working with the idea of some particular user and then for the value of that I'm going to reference

post user ID.

Now we are able to pass in post user id recall because if we go back over to the post and point every

posts object has a user id property and that is the ID of the user that made that post.

All right so now if we save this we should be able to flip back over to our application and hopefully

if you refresh the page I think that I kind of have my development server not running right now.

Believe it or not you know I definitely just change this but I'm not seeing that update so I'll tell

you what.

There we go.

All right.

That's better.

Fun see the update.

I had stopped my server between the last Wii on this one that's why I thought maybe I hadn't started

back up.

Anyways we now see the text user header here on the screen.

So that definitely looks pretty reasonable.

Now let's keep going with these are Hetter.

So I got to go back over here into the user header component.