258.2. Displaying the User Header(Section 19: Redux Store Design)
We want to make sure that we call an action creator inside of here that's going to fetch the appropriate

user.

So at the top we're going to make sure that we import the connect function and the action creator itself

and we will use the connect tag to wire up the action creator to the component.

So at the top I'm going to import connect from re-act redux and I'll import fetch user from actions.

So now down at the bottom I'll place the connect function and make sure I wrap my component with a second

set parentheses.

Right now we did not have a map state to prop's function.

I'm going to pass a null as the first argument and then for the second argument I'll do an object with

fecche user OK.

It's now the last thing we do inside of our component.

When ever this component appears on the screen we're probably going to want to make sure that we attempt

to patch the given user that this thing is supposed to show.

So I'm going to define that component did mount inside of here when whenever user header is displayed.

I'm going to call my action Creator through this stuff.

Propst fetch user.

And then remember we need to pass than the ID of the user that we want to fetch.

And as we saw just a second ago we are passing in the idea the user that we want to fetch as a prop

called user id so back over here I'll make sure I call fetch user with this stuff Propst user id OK.

And that's pretty much it.

So we can save this and flip back over to my application and it looks like everything still works.

Now we don't have any redo sir to actually pick up the data from the request that we are making and

eventually get access to it and show it on the screen.

So we'll take a quick pause right here and start to tackle that in the next video.

Now before we move on I want to remind you I told you that we were going to run into one or two little

challenges.

So if you are looking at some aspect of this application right now and saying Hey Stephen you know what

I don't think this is working the way we expect it to work.

That is fine.

I'm not going to tell you exactly what it is because I kind of want you to stumble into it yourself.

But again if you see something that looks weird that is totally fine and we're going to fix it up in

a couple of videos.

OK.

So let's take a pause right here and we'll start working on a redo in the next video.