259.1. Finding Relevant Users(Section 19: Redux Store Design)
In the last video we were able to call our fecche user action creator from our compound did mount inside

the user Hetter class.

Now in this video we're going to start to make sure that we wire up a reducer to catch that action that

is now being dispatched from our actual creator.

And we're going to make sure that this new users reduce is going to maintain a array or a simple list

of all the users that we have fetched inside of our application.

Once we collect that data inside the reducer we can then make it available to our user Hetter component

and then the user head or component can pick out that user that it cares about and shows some details

about it on the screen.

So let's get to it.

We're going to first create a new reducer by creating a file inside the reducers directory.

So I'm going to call this user's reducer.

Yes.

And then inside of your we'll put down some typical boilerplate for a producer they'll say export default

a function.

This is going to be called with arguments States and action because this producer is supposed to maintain

an array or list of records.

I'm going to make its default state value a empty array like so then inside of my producer to set up

a switch statement.

And we're going to set up one case here to watch for just the action type that this producer cares about.

If we open up the actions index file or Simien actions index yes file again you'll recall that we only

care about actions of type fetch user inside of that reducer.

So back inside of this case statement I'll say case fetch underscore user like so now if we see a reducer

or see me in action with that type right there we're going to make sure that we add the payload property

on that action to our state array.

Now we just spoke about how we can add elements to NRA member we looked at this diagram right here.

So to add an element to an array we can create a new array.

We can say dot dot dot state a comma and then the new record that we want to shove in there.

So for a case for each user I will say return dot dot dot state an action payload that is the new record

that I want to add in and we'll also set up our default case in which case we will simply return our

state OK.

That looks pretty good.

So we'll now take the seducer and make sure that we wire it up to our combined reducers call in the

reducers index such as file.

So at the top I will import users reducer from users re douceur and then we'll make sure that we add

this into the combined reducer call.

So add in a new key of users and that will get our reducer of users Regis or.

OK so not too bad.

So now last thing we really have to do here is open up our user header J.S. file.

So now inside if you're this component needs to get access to our redux level state remember any time

we want to do that we will find the Mount St. to prop's function.

So down underneath the components self I will define map state to prop's.

This will be called with the state argument.

And now we need to look at state DOT users and to find the very particular users that we care about

to get started.

Let's just return an object with the users property and will reference state DOT users.

So now our component is going to have access to a prop called this Propst users that's going to be the

array of all the users that we care about.