263.3. Memoization Issues(Section 19: Redux Store Design)(Section 19: Redux Store Design)
So let's try the second option here.

Let's try to remove the memoires out to the outer function and I want to make sure I clean up the parentheses

down there as well and then I'll try minimalizing the inner function right here.

So right in front of async I'll put underscore minimalize and I'll make sure I put my closing print

see right there in line 21 for me.

OK so now we are memorizing the interior or inner function.

This is the one that actually gets invoked with redux thunk.

So hopefully now redux sunk will only be able to call that function one time and never again after that.

All right so I'm going to save this and we'll flip back over and we'll very quickly see if we're still

making all these requests.

So even with this approach we're still running into issues and it's not doing what we expect.

So why is that.

Well remember every single time that we call this action creator we are going to declare a new function.

We recreate this function right here in memory every single time that we call fetch user.

So we recreate this function and then we memorize it and return it.

Now we definitely memorize the internal the interior function right.

So in theory it's not going to be called multiple times by redux stunk but really remember what happens

here every time we call fetch user We're going to create a new version of the interior function and

memorize it every single time.

So even though we are memorizing this thing we are memorizing a new version of the function.

Every time we call the action creator.

So every time we call the action creator new memorized version comes out and we can successfully call

that memorized version and the original interior function right here will be invoked because we are

recreating it every single time.

So like I said this problem is not quite as crystal clear as you might expect.

So now that we understand why we cannot just throw memories inside of here willy nilly.

Let's take a quick pause.

We'll take a look at how we're going to solve this problem in the next video.