264.2. One Time Memoization(Section 19: Redux Store Design)(Section 19: Redux Store Design)
Now there is one last little refactor that you could do inside of your totally optional.

If you do this refactor things start to look really crazy.

But nonetheless it is optional.

So if we wanted to we could remove this curly brace right here we could remove the curly brace and semi-colon

right there.

And then we could add new line that entire expression.

And so we would end up with something that looks like this right here.

So now we've got a function that returns a function that calls underscore fecche user with the ID and

dispatch.

So like I said this is a solution that involves just changing a couple lines of code but it has some

really crazy syntax like when you look at this thing right here in the future you're probably going

to have to really really think about it for a little bit and try to understand what's going on.

So I don't know if this is the best solution.

I would just say that this is a solution.

Now the one side effect to the solution that I want to mention is that if you ever want to re fetch

a user for some reason like if you have made a change to a user or if you know that the user data has

been updated on your API and you want to reattach it unfortunately you would not be able to do it again

using this action creator with this solution.

You can only fetch each user exactly one time inside of your application.

So if you wanted to be able to reset your user you would have to essentially like declare a another

action creator that has the same logic but just does not apply that memoization step.

So like I said maybe not the best solution but it is a solution that fixes our problem.

All right.

Let's take a pause right here in the next video.

I'm going to show you a second solution to this problem.

That's a little bit more universal I suppose.

Or it's going to work out in more cases a little bit better because we are not having to do this memoization

step.

However it's going to involve changing a lot more code than what we had to do for this right here.

So let's take a quick pause and we'll take a look at that solution in the next video.